+++ CHANGELOG.md
@@ -4,11 +4,29 @@ All notable changes to `@todovue/tv-theme-button` will be documented in this fil
 
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+---
+## [1.0.3] - 2025-11-15
+### 🛠️ Changed
+- Styles are now served as a separate CSS file (`dist/tv-theme-button.css`) generated by Vite during the build process.
+- Users must now explicitly import the stylesheet in their application:
+  - For Vue/Vite SPA: `import '@todovue/tv-theme-button/style.css'` in `main.ts`
+  - For Nuxt 3/4: Add `'@todovue/tv-theme-button/style.css'` to the `css` array in `nuxt.config.ts`
+
+### ✨ Added
+- Added `sideEffects` field to `package.json` to support proper tree-shaking with CSS files.
+- Added `./style.css` export path in `package.json` for explicit CSS imports.
+- Improved SSR/SSG compatibility, especially for Nuxt 3/4 applications.
+
+### 📦 Dependencies
+- Removed dependency on `vite-plugin-css-injected-by-js` from `devDependencies`
+
+---
 ## [1.0.2] - 2025-11-13
 ### 🐛 Fixed
 - Fixed SSR hydration error in Nuxt caused by localStorage access during server-side rendering
 - Fixed component import path in demo application
 
+---
 ## [1.0.1] - 2025-10-19
 ### 🛠️ Changed
 - The library build now uses `src/entry.ts` (exports both the component and the plugin) instead of directly exporting the `.vue` file.
@@ -20,8 +38,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Explicit export of the style file: `import '@todovue/tv-theme-button/style.css'`.
 - Documentation for usage in SSR and Nuxt 3 applications.
 
+---
 ## [1.0.0] - 2025-05-05
-
 ### ✨ Added
 
 - Initial release of `TvThemeButton`
