+++ src/demo/Demo.vue
@@ -1,9 +1,8 @@
 <script setup>
-import { TvDemo } from '@todovue/tv-demo';
 import { defineAsyncComponent } from 'vue';
 
 const TvThemeButton = defineAsyncComponent(() => import('../components/TvThemeButton.vue'));
-import { demos } from "../utils/mocks.js";
+import { demos } from './utils/mocks.js';
 </script>
 
 <template>
@@ -14,6 +13,6 @@ import { demos } from "../utils/mocks.js";
     npm-install="@todovue/tv-theme-button"
     source-link="https://github.com/TODOvue/tv-theme-button"
     url-clone="https://github.com/TODOvue/tv-theme-button.git"
-    version="1.0.2"
+    version="1.0.3"
   />
 </template>
