+++ src/entry.ts
@@ -1,4 +1,5 @@
 import TvThemeButton from './components/TvThemeButton.vue'
+import './style.scss'
 
 (TvThemeButton as any).install = (app: any) => {
     app.component('TvThemeButton', TvThemeButton)
